+++ test/lightning/invocation/log_line_test.exs
@@ -1,0 +1,55 @@
+defmodule Lightning.Invocation.LogLineTest do
+  use Lightning.DataCase, async: true
+
+  import Lightning.Factories
+
+  alias Lightning.Invocation.LogLine
+
+  test "new/2" do
+    dataclip = insert(:dataclip)
+
+    %{triggers: [trigger]} = workflow = insert(:simple_workflow)
+
+    work_order =
+      insert(:workorder,
+        workflow: workflow,
+        trigger: trigger,
+        dataclip: dataclip
+      )
+
+    run =
+      insert(:run,
+        work_order: work_order,
+        starting_trigger: trigger,
+        dataclip: dataclip
+      )
+
+    log_line =
+      LogLine.new(
+        run,
+        %{message: "Hello, World!", timestamp: DateTime.utc_now()},
+        nil
+      )
+
+    assert log_line.valid?
+
+    log_line =
+      LogLine.new(
+        run,
+        %{message: "", timestamp: DateTime.utc_now()},
+        nil
+      )
+
+    assert log_line.valid?, "should be able to have an empty message"
+
+    log_line =
+      LogLine.new(
+        run,
+        %{message: nil, timestamp: DateTime.utc_now()},
+        nil
+      )
+
+    assert {:message, {"can't be blank", []}} in log_line.errors
+    refute log_line.valid?
+  end
+end